backend/requirements_prod.txt
# Production Requirements for eBay Listing Optimizer Backend
# Core Framework
fastapi==0.104.1
uvicorn[standard]==0.24.0
gunicorn==21.2.0

# Database & ORM
sqlalchemy==2.0.23
alembic==1.13.0

# Authentication & Security
python-jose[cryptography]==3.3.0
python-multipart==0.0.6
passlib[bcrypt]==1.7.4
bcrypt==4.1.2

# Pydantic & Validation
pydantic==2.5.0
pydantic-settings==2.1.0
email-validator==2.1.0

# Google Services
google-auth==2.24.0
google-auth-oauthlib==1.1.0
google-auth-httplib2==0.1.1
google-api-python-client==2.108.0
gspread==5.12.0

# HTTP & Networking  
httpx==0.25.2
requests==2.31.0

# Utilities & Performance
python-dotenv==1.0.0
redis==5.0.1
celery==5.3.4

# Monitoring & Logging
structlog==23.2.0
sentry-sdk[fastapi]==1.40.0

# Development & Testing (minimal for production debugging)
pytest==7.4.3
pytest-asyncio==0.21.1

# Data Processing
pandas==2.1.3
numpy==1.25.2

# Production Server
uvloop==0.19.0
httptools==0.6.1